+++ atable/CHANGELOG.md
@@ -1,6 +1,25 @@
 # Change Log - @stonecrop/atable
 
-This log was last generated on Fri, 14 Aug 2026 08:05:46 GMT and should not be manually modified.
+This log was last generated on Mon, 17 Aug 2026 19:08:23 GMT and should not be manually modified.
+
+## 0.21.0
+Mon, 17 Aug 2026 19:08:23 GMT
+
+### Patches
+
+- Render link cells from { id, displayText } objects returned by native queries instead of fieldname__display siblings.
+
+## 0.20.0
+Mon, 17 Aug 2026 18:41:19 GMT
+
+_Version update only_
+
+## 0.19.0
+Mon, 17 Aug 2026 12:46:25 GMT
+
+### Minor changes
+
+- Use a link's pre-resolved display text from the row instead of refetching it.
 
 ## 0.18.0
 Fri, 14 Aug 2026 08:05:46 GMT