+++ themes/CHANGELOG.md
@@ -1,6 +1,21 @@
 # Change Log - @stonecrop/themes
 
-This log was last generated on Fri, 14 Aug 2026 08:05:46 GMT and should not be manually modified.
+This log was last generated on Mon, 17 Aug 2026 19:08:23 GMT and should not be manually modified.
+
+## 0.21.0
+Mon, 17 Aug 2026 19:08:23 GMT
+
+_Version update only_
+
+## 0.20.0
+Mon, 17 Aug 2026 18:41:19 GMT
+
+_Version update only_
+
+## 0.19.0
+Mon, 17 Aug 2026 12:46:25 GMT
+
+_Version update only_
 
 ## 0.18.0
 Fri, 14 Aug 2026 08:05:46 GMT